Refactoring - Comply with PSR-4 to support NC >= 27
